1. 发现严重BUG：串口未连接下载器时特定函数Str_check_test不生效。 2. 使用了新的字符串对比函数Str_check_test，对比速度更快，使用更简便 3. 简单修改了read_date发送数据写法，实际发送数据无差别 4. 主函数目前为测试状态

+++ user/uart.c
@@ -68,7 +68,7 @@ unsigned char Str_check(char *str, char *con)
     return 0;
 }
 
-unsigned char Str_check_test(char *str, char *con, uint8 number)
+unsigned char Str_check_test(char *str, char *con, unsigned int number)
 {
     unsigned int i   = 0;
     unsigned int cnt = 0;
